+++ tests/ops/conftest.py
@@ -46,6 +46,7 @@ from .fixtures.saas.connection_template_fixtures import *
 from .fixtures.saas.datadog_fixtures import *
 from .fixtures.saas.hubspot_fixtures import *
 from .fixtures.saas.mailchimp_fixtures import *
+from .fixtures.saas.marketo_fixtures import *
 from .fixtures.saas.outreach_fixtures import *
 from .fixtures.saas.request_override.mailchimp_override_fixtures import *
 from .fixtures.saas.rollbar_fixtures import *